+++ src/__tests__/protected-promise.test.js
@@ -3,51 +3,31 @@ const {
   PromiseProtectedRejected,
 } = require('../protected-promise')
 
-const BluebirdPromise = require('bluebird')
-
 describe('ProtectedPromise', () => {
   test('protectPromise(simple promise, resolves)', async () => {
-    const thePromise = Promise.resolve(123)
-    expect(thePromise).not.toBeInstanceOf(BluebirdPromise)
+    const thePromise = protectPromise(Promise.resolve(123))
     expect(thePromise).toBeInstanceOf(Promise)
 
-    const theProtectedPromise = protectPromise(thePromise)
-
-    expect(theProtectedPromise).resolves.toBe(123)
+    expect(thePromise).resolves.toBe(123)
   })
 
   test('protectPromise(simple promise, rejects)', async () => {
-    const thePromise = Promise.resolve().then(() => {
-      throw Error('TEST')
-    })
-    const theProtectedPromise = protectPromise(thePromise)
-
-    expect(theProtectedPromise).resolves.toBeInstanceOf(
-      PromiseProtectedRejected,
+    const thePromise = protectPromise(
+      Promise.resolve().then(() => {
+        throw Error('TEST')
+      }),
     )
+
+    expect(thePromise).resolves.toBeInstanceOf(PromiseProtectedRejected)
   })
 
-  test('protectPromise(bluebird promise, resolves)', async () => {
-    const thePromise = BluebirdPromise.resolve(123)
-    expect(thePromise).toBeInstanceOf(BluebirdPromise)
-    expect(thePromise).not.toBeInstanceOf(Promise)
-
-    const theProtectedPromise = protectPromise(thePromise)
-
-    expect(theProtectedPromise).resolves.toBe(123)
-  })
-
-  test('protectPromise(bluebird promise, rejects)', async () => {
-    const thePromise = BluebirdPromise.resolve().then(() => {
-      throw Error('TEST')
-    })
-    expect(thePromise).toBeInstanceOf(BluebirdPromise)
-    expect(thePromise).not.toBeInstanceOf(Promise)
-
-    const theProtectedPromise = protectPromise(thePromise)
-
-    expect(theProtectedPromise).resolves.toBeInstanceOf(
-      PromiseProtectedRejected,
+  test('protectPromise(promise, rejects)', async () => {
+    const thePromise = protectPromise(
+      Promise.resolve().then(() => Promise.reject(Error('TEST'))),
     )
+    expect(thePromise).toBeInstanceOf(Promise)
+    expect(thePromise).not.toBeInstanceOf(Error)
+
+    expect(thePromise).resolves.toBeInstanceOf(PromiseProtectedRejected)
   })
 })